[#14]refactor: login, signup 패키지 합침
