--- src/test/java/soma/haeya/edupi_user/signup/SignUpControllerTest.java
@@ -1,18 +1,0 @@
-package soma.haeya.edupi_user.signup;
-
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.web.servlet.MockMvc;
-
-@AutoConfigureMockMvc
-@SpringBootTest
-class SignUpControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-
-}
